[Change] Re-organized project structure
