+++ frontend/src/components/posts/PostContentEditor.jsx
