+++ frontend/src/pages/Platforms.module.css
